Fixed GUI launch
